+++ src/screens/BlueDuckV/sections/Section1/HeroSection.jsx
